move dbcontext to data namespace
