+++ Areas/Identity/IdentityHostingStartup.cs
@@ -1,6 +1,6 @@
 using System;
 using Cribbly.Models;
-using Cribbly.Services;
+using Cribbly.Data;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Identity.UI;
