+++ Data/AppDbContext.cs
@@ -3,7 +3,7 @@ using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
 
-namespace Cribbly.Services
+namespace Cribbly.Data
 {
     public class AppDbContext : IdentityDbContext<Player>
     {
@@ -19,6 +19,13 @@ namespace Cribbly.Services
             modelBuilder.Entity<Player>()
                 .HasOne<Team>(p => p.Team)
                 .WithMany(t => t.Players);
+
+            modelBuilder.Entity<TeamPlayInGame>()
+                .HasKey(t => new { t.TeamID, t.PlayInGameID });
+
+            modelBuilder.Entity<TeamPlayInGame>()
+                .HasOne(tg => tg.Team)
+                .WithMany(t => )
         }
 
         public DbSet<Team> Teams { get; set; }
